Move c-api helper functions to header for reuse
